separated react components from @gateway/theme (#433)
