+++ libs/theme-react/.babelrc
