+++ libs/theme-react/README.md
@@ -1,0 +1,7 @@
+# theme-react
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test theme-react` to execute the unit tests via [Jest](https://jestjs.io).
