+++ libs/theme-react/jest.config.ts
@@ -1,0 +1,9 @@
+module.exports = {
+  displayName: 'theme-react',
+  preset: '../../jest.preset.ts',
+  transform: {
+    '^.+\\.[tj]sx?$': 'babel-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../coverage/libs/theme-react',
+};
