+++ libs/theme-react/project.json
@@ -1,0 +1,23 @@
+{
+  "root": "libs/theme-react",
+  "sourceRoot": "libs/theme-react/src",
+  "projectType": "library",
+  "tags": [],
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/theme-react/**/*.{ts,tsx,js,jsx}"]
+      }
+    },
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["coverage/libs/theme-react"],
+      "options": {
+        "jestConfig": "libs/theme-react/jest.config.ts",
+        "passWithNoTests": true
+      }
+    }
+  }
+}
