+++ libs/theme-react/src/lib/theme-provider/theme-provider.tsx
@@ -1,9 +1,10 @@
 import { PropsWithChildren } from 'react';
 
+import { theme } from '@gateway/theme';
+
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider as MUIThemeProvider } from '@mui/material/styles';
 
-import { theme } from '../../lib/theme';
 import { GlobalStyles } from './global-styles';
 
 type Props = {
