+++ libs/theme/project.json
@@ -4,6 +4,16 @@
   "projectType": "library",
   "tags": [],
   "targets": {
+    "build": {
+      "executor": "@nrwl/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/theme",
+        "main": "libs/theme/src/index.ts",
+        "tsConfig": "libs/theme/tsconfig.lib.json",
+        "assets": ["libs/theme/*.md"]
+      }
+    },
     "lint": {
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
