+++ libs/theme/src/index.ts
@@ -1,6 +1,5 @@
-export * from './ui/theme-provider/theme-provider';
 export * from './lib/theme';
 export * from './lib/types';
-export * as TOKENS from './ui/style-tokens/style-tokens';
+export * as TOKENS from './lib/style-tokens';
 export * from './types';
 export { default as brandColors } from './lib/config/colors';
