+++ package.json
@@ -30,6 +30,7 @@
     "@nrwl/next": "14.0.3",
     "@prisma/client": "^4.7.1",
     "@rainbow-me/rainbowkit": "^0.8.1",
+    "@react-email/components": "^0.0.3",
     "@solana/wallet-adapter-base": "^0.9.20",
     "@solana/wallet-adapter-react": "^0.15.28",
     "@solana/wallet-adapter-react-ui": "^0.9.27",
@@ -73,6 +74,7 @@
     "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "18.0.0",
     "react-easy-crop": "^4.2.0",
+    "react-email": "^1.8.0-canary.1",
     "react-hook-form": "^7.31.1",
     "react-icons": "^4.4.0",
     "react-json-view": "^1.21.3",
@@ -100,8 +102,8 @@
     "@graphql-codegen/import-types-preset": "^2.2.6",
     "@graphql-codegen/typescript": "^3.0.1",
     "@graphql-codegen/typescript-graphql-request": "^4.5.8",
-    "@graphql-codegen/typescript-resolvers": "^3.1.0",
     "@graphql-codegen/typescript-operations": "^3.0.1",
+    "@graphql-codegen/typescript-resolvers": "^3.1.0",
     "@nrwl/cli": "14.0.3",
     "@nrwl/cypress": "14.0.3",
     "@nrwl/eslint-plugin-nx": "14.0.3",
@@ -135,18 +137,18 @@
     "dotenv": "^16.0.0",
     "eslint": "~8.12.0",
     "eslint-config-next": "12.1.5",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-cypress": "^2.10.3",
     "eslint-plugin-editorconfig": "^3.2.0",
     "eslint-plugin-import": "2.25.4",
     "eslint-plugin-import-helpers": "^1.2.1",
     "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "7.29.4",
     "eslint-plugin-react-hooks": "4.4.0",
     "jest": "27.5.1",
     "nx": "14.0.3",
-    "prettier": "^2.5.1",
+    "prettier": "^2.8.5",
     "prisma": "^4.7.1",
     "react-docgen-typescript": "^2.2.2",
     "react-test-renderer": "18.0.0",
